Move in_tail_format.js under app/javascript/packs
