+++ app/views/shared/vue/_in_tail_format.html.erb
@@ -1,27 +1,31 @@
-<script type="text/template" id="format-bundle">
+<%- add_javascript_pack_tag("in_tail_format") %>
+
+<script type="text/x-template" id="format-bundle">
+  <div>
     <div class="form-inline form-group">
       <label for="in_tail_setting_format">format</label>
       <select id="in_tail_setting_format" name="setting[format]" v-model="format" class="form-control">
-        <option v-repeat="selectableFormats" value="{{ $value }}" v-attr="selected: format==$value">{{ $value }}</option>
+        <option v-for="selectableFormat in selectableFormats" v-bind:value="selectableFormat" v-bind:selected="format==selectableFormat" v-if="selectableFormats">{{ selectableFormat }}</option>
       </select>
     </div>
-    <div class="form-inline form-group" v-repeat="options">
-      <label for="in_tail_setting_{{ $value }}" v-if="!useTextArea">{{ $value }} </label>
-      <input id="in_tail_setting_{{ $value }}" type="{{ useTextArea ? 'hidden' : 'text' }}" name="setting[{{ $value }}]" v-model="params.setting[$value]" v-on="keyup: onKeyup" size="100%" class="form-control" />
+    <div class="form-inline form-group" v-for="option in options">
+      <label v-bind:for="'in_tail_setting_' + option" v-if="!useTextArea">{{ option }} </label>
+      <input v-bind:id="'in_tail_setting_' + option" v-bind:type="useTextArea ? 'hidden' : 'text'" v-bind:name="'setting[' + option + ']'" v-model="params.setting[option]" v-on:keyup="onKeyup" size="100%" class="form-control" />
     </div>
 
     <div v-if="useTextArea">
       <div class="form-inline form-group">
         <label for="in_tail_setting_format_firstline">format_firstline</label>
-        <input id="in_tail_setting_format_firstline" type="text" name="setting[format_firstline]" v-model="params.setting['format_firstline']" v-on="keyup: onKeyup" size="100%" class="form-control" />
+        <input id="in_tail_setting_format_firstline" type="text" name="setting[format_firstline]" v-model="params.setting['format_firstline']" v-on:keyup="onKeyup" size="100%" class="form-control" />
       </div>
 
       <div class="form-group">
        <p class="alert alert-warning"><%= t("fluentd.settings.in_tail.notice_for_multiline_limit") %></p>
         <label for="in_tail_setting_formats">formats</label>
-        <textarea id="in_tail_setting_formats" type="text" name="setting[formats]" v-model="params.setting['formats']" v-on="keyup: onKeyup" rows='20' size='100%' class="form-control"></textarea>
+        <textarea id="in_tail_setting_formats" type="text" name="setting[formats]" v-model="params.setting['formats']" v-on:keyup="onKeyup" rows='20' size='100%' class="form-control"></textarea>
       </div>
     </div>
+  </div>
 </script>
 
 <div id="in_tail_format"
@@ -31,9 +35,13 @@
   paramsJson="<%= params.to_json %>"
   >
 
-  <div v-component="format-bundle" wait-for="data-loaded"></div>
+  <format-bundle
+    v-bind:format-options="formatOptions"
+    v-bind:format="format"
+    v-bind:params="params"
+    v-on:update-preview="updatePreview"></format-bundle>
 
-  <pre>{{{ highlightedLines }}}</pre>
+  <pre>{{ highlightedLines }}</pre>
 
   <div class="well well-sm">
     <%= raw t('fluentd.settings.in_tail_option_guide') %>
